+++ src/Qualys/functions/public/Add-QualysUser.ps1
@@ -5,7 +5,7 @@
     Adds a new user to Qualys
 .PARAMETER Credential
     This API call only supports basic HTTP authentication. You must provide your credentials separately for this function.
-.PARAMETER ExternalID
+.PARAMETER External_ID
     Set a custom External ID (required for SSO)
 .PARAMETER SendEmail
     Specifies whether the new user will receive an email notification with a secure link to their login credentials.
@@ -52,6 +52,8 @@
 function Add-QualysUser{
     [CmdletBinding()]
     param (
+        [Parameter(Mandatory=$true)]
+        [System.Management.Automation.PSCredential]$Credential,
         [Switch]$SendEmail,
         [Parameter(Mandatory=$true)]
         [String]$Role,
@@ -74,9 +76,8 @@ function Add-QualysUser{
         [Parameter(Mandatory=$true)]
         [String]$Country,
         [String]$State,
-        [String]$ExternalID,
-        [Parameter(Mandatory=$true)]
-        [System.Management.Automation.PSCredential]$Credential
+        [String]$External_ID
+
     )
 
     process{
@@ -99,7 +100,7 @@ function Add-QualysUser{
                 city = $City
                 country = $Country
                 state = $State
-                external_id = $ExternalID
+                external_id = $External_ID
             }
         }