+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.1.1] - 2020-11-23
+
+### Added
+
+-
+
+### Changed
+
+- Update User functions ExternalID to match API External_ID
+- Made Credential first parameter of User functions
+
 ## [1.1.0] - 2020-11-17
 
 ### Added